stop pruning "will always be core" libraries
1.  They might remain core forever, but they weren't always core.  For
example, prior to this commit we skipped feature.pm, even though requiring it
means something -- right now, for example, it would require 5.10 -- but what
if feature.pm is backported to 5.8.20?  The same goes for warnings.pm

2.  They might not remain in core forever!  This is why I already objected to
using Module::CoreList.  The cost of including "strict" in prereqs seems
miniscule.  At worst, it means people have to skim a few more lines here and
there on the rare occasions when they have to read META.json files.

3.  The alternative is to keep writing more and more machinery to figure out
what stuff is safe to exclude.  This seems pointless.

4.  BUT!!  We must omit lib.pm, because it does not appear in 02packages,
and I think had we better avoid requiring things that can't be found in that
index, even if they *are* pretty much guaranteed to be around forever.

Thanks to mauke for bringing this up.